Move synonym tests into topical classes; drop redundant Name override (#228 review)
Address PR review feedback:
- Relocate the synonym/vocabulary assertions out of the dedicated
  SynonymVocabularyTests classes into the topical homes for each item:
  - Zork: map "parchment" and leaflet mail/booklet/pamphlet/advertisement go to
    MiscItemTests (behavioral [TestCase] rows); skeleton bones/body join its
    CannotBeTaken cases; "examine dam"/"examine gates" goes to DamTests.
  - Planetfall: the mutant "monster" synonym test goes to BioLabEscapeTests;
    lamp "lantern" to LabsAndDoorsTests (beside RadLab_HaveLamp); computer-output
    "printout" to ComputerRoomTests.
  Both SynonymVocabularyTests classes are removed.
- Drop the redundant FloodControlDam.Name override: ItemBase.Name already
  defaults to NounsForMatching.First() ("dam").
